+++ src/test/java/es/upm/grise/profundizacion/HandleDocuments/DocumentIdProviderFake.java
@@ -1,0 +1,10 @@
+package es.upm.grise.profundizacion.HandleDocuments;
+
+public class DocumentIdProviderFake implements DocumentIdProvider {
+
+	@Override
+	public int getDocumentId() throws NonRecoverableError {
+		return 1;
+	}
+
+}
